- integration of installation files for remaining external software     - trust.sh     - trf.sh     - treks.sh (their website down, automated download failes till now) - integrated option uninstall_all_ext_software.sh into deleteTRAL.sh - changed in all scripts copy without overwriting (cp -n) to copy with overwrite (cp)     reason: update files in systempath after change in script or in paths - setupTRAL.sh: integrated function to install all external software which you like - included test to see wheter a ext software already installed - added some comments and structure to some scripts - created first README for easy_setup

+++ easy_setup/install_ext_software/phobos.sh
@@ -28,40 +28,47 @@ PARENT_PATH=$( cd "$(dirname "${BASH_SOURCE[0]}")" ; cd .. ; pwd -P )
 ######################
 ### Download and Installion PHOBOS
 
-echo -e  "\nThis program Phobos is copyright protected. It is only distributed from the authors web page (www.rub.de/spezzoo/cm/cm_phobos.htm).\n"
-echo -e  "\nWould you like to register and download the program manually? A popup will open."
-echo -e  "Otherwise you can directly download the phobos-v3.3.12-linux version automatically.\n"
-read -p "manually (m) or automatically (a)?:" ma
-case $ma in
-    [Mm]* )                                                                 # redirects user to download page
-        if which xdg-open > /dev/null
-        then
-            xdg-open https://www.ruhr-uni-bochum.de/ecoevo/cm/regist_form.htm
-        elif which gnome-open > /dev/null
-        then
-            gnome-open https://www.ruhr-uni-bochum.de/ecoevo/cm/regist_form.htm
-        fi
+if [ ! -f phobos_* ]; then # test if not already in directory
 
-        echo -e  "\nAfter downloading your version of choice you can unzip it and put the binaries into your PATH.\n"
-        ;;
-    [Aa]* ) 
-        echo -e "\nDo you confirm to the authors copyright (Copyright (c) Christoph Mayer 2006-2017)?\n This program is for academic and non-commercial usage only."
-        read -p "Would you like to download the phobos-v3.3.12-linux version? yes(y) or no (n):" yn
-        case $yn in
-            [Yy]* )                                                            
-                LINK_PHOBOS=http://www.rub.de/ecoevo/cm/phobos-v3.3.12-linux.tar.gz 
-                wget $LINK_PHOBOS -P $TRAL_EXT_SOFTWARE
-                tar zxf $TRAL_EXT_SOFTWARE/phobos-v3.3.12-linux.tar.gz -C $TRAL_EXT_SOFTWARE
-                rm -rf $TRAL_EXT_SOFTWARE/phobos-v3.3.12-linux.tar.gz
-                cp -n $TRAL_EXT_SOFTWARE/phobos-v3.3.12-linux/bin/phobos_* /usr/local/bin/      # copies binaries to system path /usr/local/bin/
-                echo -e "\nPHOBOS binaries are now in the user path /usr/local/bin/"
-                ;;
-            [Nn]* ) 
-                echo -e "Abort."
-                ;;
-        esac
-        ;;
-esac
+    echo -e  "\nThis program Phobos is copyright protected. It is only distributed from the authors web page (www.rub.de/spezzoo/cm/cm_phobos.htm).\n"
+    echo -e  "\nWould you like to register and download the program manually? A popup will open."
+    echo -e  "Otherwise you can directly download the phobos-v3.3.12-linux version automatically.\n"
+    read -p "manually (m) or automatically (a)?:" ma
+    case $ma in
+        [Mm]* )                                                                 # redirects user to download page
+            if which xdg-open > /dev/null
+            then
+                xdg-open https://www.ruhr-uni-bochum.de/ecoevo/cm/regist_form.htm
+            elif which gnome-open > /dev/null
+            then
+                gnome-open https://www.ruhr-uni-bochum.de/ecoevo/cm/regist_form.htm
+            fi
+
+            echo -e  "\nAfter downloading your version of choice you can unzip it and put the binaries into your PATH.\n"
+            ;;
+        [Aa]* ) 
+            echo -e "\nDo you confirm to the authors copyright (Copyright (c) Christoph Mayer 2006-2017)?\n This program is for academic and non-commercial usage only."
+            read -p "Would you like to download the phobos-v3.3.12-linux version? yes(y) or no (n):" yn
+            case $yn in
+                [Yy]* )                                                             
+                    LINK_PHOBOS=http://www.rub.de/ecoevo/cm/phobos-v3.3.12-linux.tar.gz 
+                    wget $LINK_PHOBOS -P $TRAL_EXT_SOFTWARE
+                    tar zxf $TRAL_EXT_SOFTWARE/phobos-v3.3.12-linux.tar.gz -C $TRAL_EXT_SOFTWARE
+                    rm -rf $TRAL_EXT_SOFTWARE/phobos-v3.3.12-linux.tar.gz
+                    cp $TRAL_EXT_SOFTWARE/phobos-v3.3.12-linux/bin/phobos_* /usr/local/bin/      # copies binaries to system path /usr/local/bin/
+                    echo -e "\nPHOBOS binaries are now in the user path /usr/local/bin/"
+                    ;;
+                [Nn]* ) 
+                    echo -e "Abort."
+                    ;;
+            esac
+            ;;
+    esac
+
+    else
+    echo "PHOBOS already installed." && exit 0
+
+fi
 
 
 # Per default phobos is accessible via phobos_* (as long as the binary is copied to the system path).
